Only verify text segment since .got.plt will change during load
During program startup, some of the function pointers in the .got.plt
section will be modified, causing a checksum computed over the
containing segment to differ from a checksum produced before the file
was loaded. Only verify the text segment to work around this.

The checksums were also moved to the .text section so that competitors
can see how we are checksumming "around" the values.