Remove the second checksum to make challenge easier
Since kernel.c never actually verifies the second checksum, it might
take a few guesses for competitors to discover that the second
checksum is of the entire data segment. To make things easier and
hopefully get more solves, only use one checksum that is clearly
computed in kernel.c.

To use 2 checksums, revert this commit.